chore(refactor): move config to an internal package
